Add files via upload
This was built as a way for me to look into grabbing values from inputs and creating applications that respond to that user input. This is not a practical use application, as many local areas have different tax rates than the state's base rate, but does show the functionality and potential of a similar fully functional application.